modules/2_patch_generation/requirements.txt
torch>=1.12.0
torchvision>=0.13.0
Pillow>=9.0.0
